[IMP] 将32位clusterId修改为64位

See merge request hzero-c7ncd/choerodon-cluster-agent!59

+++ pkg/websocket/client.go
@@ -54,7 +54,7 @@ type appClient struct {
 	backgroundWait sync.WaitGroup
 	pipeConns      map[string]*websocket.Conn
 	respQueue      []*model.Packet
-	clusterId      int32
+	clusterId      int64
 	gitRepos       map[string]*git.Repo
 }
 
@@ -64,7 +64,7 @@ func NewClient(
 	t Token,
 	endpoint string,
 	crChannel *channel.CRChan,
-	clusterId int32,
+	clusterId int64,
 	gitRepos map[string]*git.Repo) (Client, error) {
 	if endpoint == "" {
 		return nil, fmt.Errorf("no upstream URL given")
